Address Copilot review comments on video player tests
- Extract shared reduced-motion helpers and shell commands into VideoPlayerTestRule (ExternalResource) to eliminate duplication between BpkVideoPlayerTest and BpkVideoPlayerDefaultControlsTest
- Fix dispose test to wait for Playing state before removing the player from composition, making the STATE_IDLE assertion unambiguously prove release() was called
